Cleaned up compute.glsl. Now works with multiple layers
All I needed for multiple layers was to divide the cost when adjusting the previous layers new expected value, by 10, 20, or just some whole number above 1 (where lower numbers are more coarse but quicker training, and higher numbers results in finer but slower training). The reason for this is that after 10, 20, or however many backpropagations, the result is the same as all those backpropagations will averaged out.