Learned and implemented proper backpropagation calculus
Now uses the chain rule (as it's supposed to) in order to backpropagate. Previously, only 1-2 layers were stable before errors grew. The backprop1 algorithm that was previously used does not approach the correct value.

This proper backpropagation requires much more training time, but it will approach the correct value (tested with 3 hidden layers).