New design can now perform the same previous autoencoder tasks with an ANN.
The new design allows layers that can be dynamically pushed at runtime, without the need to actually compile or mark the network as "prepared". In theory, the layers could be changed per frame, while keeping any layer you don't remove.

Need to fix the ability to save the network. In addition, It would be nice to specify which layers to include in the save (default=all). This could let you train a single autoencoder, then swap out the tail end from a loaded autodecoder to change functionality. Allows backpropagating for each type of network on the same encoder-end.